move contract-details to view
