+++ src/view/ContractDetailView/contract-detail.tsx
